kafka_producer/requirements.txt
APScheduler==3.11.0
tzlocal==5.2
kafka-python-ng==2.2.3
psycopg2-binary==2.9.10